Addressed comments on template refactoring.
