+++ gobblin-api/src/main/java/gobblin/runtime/api/Spec.java
@@ -10,7 +10,7 @@
  * CONDITIONS OF ANY KIND, either express or implied.
  */
 
-package gobblin.runtime;
+package gobblin.runtime.api;
 
 import java.net.URI;
